apps/image_optimizer/requirements.txt
Pillow==10.3.0
-e ../../libs/image_processing
